Add a parameter to allow the bypass of the BWT post processing steps: MTFT and ZLT.
In other words, it is now possible to run a raw BWT now.

Even though the BWT is a transform (input size == output size), the codec without ZLT is still a function (output size != input size in general).

The output buffer must be bigger than the input buffer since the BWT primary index must be stored in the output (up to 4 bytes). See getMaxEncodedLength().